functions-python/create_match_link/requirements.txt
firebase-admin==5.2.0
protobuf==3.20.1
py-firebase-dynamic-links==1.0.1
